댓글 삭제 시 텍스트 변경 되도록 수정, 채팅 이미지 전송 시 커스텀 모달 선택 박스 생성, 라커룸 수정 모달 null 오류 조건 추가, 커스텀 앱바 뒤로가기 버튼 추가

+++ lib/screen/subScreen/playHistory/play_history_comment_item.dart
@@ -91,7 +91,7 @@ class _PlayHistoryCommentItemState extends State<PlayHistoryCommentItem> {
                         ),
                       ),
 
-                      if (widget.matchHistoryCommentModel.userId == widget.loginMemberId)
+                      if (widget.matchHistoryCommentModel.userId == widget.loginMemberId && widget.matchHistoryCommentModel.useYn == "Y")
                         GestureDetector(
                           onTap: (){
                             widget.deleteCommentCallBack(widget.matchHistoryCommentModel);
@@ -127,9 +127,9 @@ class _PlayHistoryCommentItemState extends State<PlayHistoryCommentItem> {
                           ),
                         ),
                         Text(
-                          widget.matchHistoryCommentModel.content.toString().split(' ')[1] ?? '',
+                          widget.matchHistoryCommentModel.useYn == "N" ? "삭제된 댓글입니다" : widget.matchHistoryCommentModel.content.toString().split(' ')[1] ?? '',
                           style: TextStyle(
-                            color: Colors.white,
+                            color: widget.matchHistoryCommentModel.useYn == "N" ? Colors.grey : Colors.white,
                             fontWeight: FontWeight.w700,
                             fontFamily: 'EHSMB',
                           ),
@@ -143,9 +143,9 @@ class _PlayHistoryCommentItemState extends State<PlayHistoryCommentItem> {
                       spacing: 6,
                       children: [
                         Text(
-                          widget.matchHistoryCommentModel.content ?? '',
+                          widget.matchHistoryCommentModel.useYn == "N" ? "삭제된 댓글입니다" : widget.matchHistoryCommentModel.content ?? '',
                           style: TextStyle(
-                              color: Colors.white,
+                              color: widget.matchHistoryCommentModel.useYn == "N" ? Colors.grey : Colors.white,
                               fontWeight: FontWeight.w700,
                               fontFamily: 'EHSMB'
